선생님 시험 기능 완성, 학생 시험 중

+++ src/main/java/kr/co/codemaker/teacher/course/exam/controller/ExamController.java
@@ -6,11 +6,8 @@ import java.util.List;
 import javax.annotation.Resource;
 import javax.servlet.http.HttpSession;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
@@ -23,7 +20,6 @@ import kr.co.codemaker.teacher.course.exam.vo.ExamVO;
 import kr.co.codemaker.teacher.course.exam.vo.LessonVO;
 import kr.co.codemaker.teacher.course.exam.vo.QuestionVO;
 import kr.co.codemaker.teacher.course.exam.vo.SubjectVO;
-import kr.co.codemaker.teacher.course.lesson.service.LessonIndexService;
 
 /**
  * 
@@ -50,11 +46,6 @@ public class ExamController {
 	@Resource(name = "answersheetService")
 	private AnswersheetService answersheetService;
 
-	@Resource(name = "lessonIndexService")
-	private LessonIndexService lessonIndexService;
-	
-	private static final Logger logger = LoggerFactory.getLogger(ExamController.class);
-	
 	/**
 	 * 시험 리스트 조회 - 과목 조회, 강의 조회, 시험 조회
 	 * 
@@ -146,41 +137,9 @@ public class ExamController {
 		model.addAttribute("questionList", questionList);
 		model.addAttribute("answersheetLists", answersheetLists);
 
-		return "teacherPage/teacher/exam/examSelect";
-	}
-	
-	/**
-	 * 시험 수정화면을 요청하는 메서드
-	 * 
-	 * @author 김미연
-	 * @return
-	 */
-	@RequestMapping(path = "/exam/updateViewExam")
-	public String updateViewExam(ExamVO examVO, Model model) {
-		ExamVO ev = new ExamVO();
-		List<QuestionVO> questionList = new ArrayList<>();
-		List<AnswersheetVO> answersheetLists = new ArrayList<>();
-		
-		try {
-			ev = examService.selectExam(examVO);
-			questionList = questionService.selectQuestion(examVO);
-			for (QuestionVO questionVO : questionList) {
-				List<AnswersheetVO> answersheetList = answersheetService.selectAnswersheet(questionVO);
-				
-				for (AnswersheetVO answersheetVO : answersheetList) {
-					answersheetLists.add(answersheetVO);
-				}
-			}
-		} catch (Exception e) {
-			e.printStackTrace();
-		}
-		model.addAttribute("ev", ev);
-		model.addAttribute("questionList", questionList);
-		model.addAttribute("answersheetLists", answersheetLists);
-
 		return "teacherPage/teacher/exam/examUpdate";
 	}
-
+	
 	/**
 	 * 시험을 수정하는 메서드
 	 * 
@@ -189,6 +148,7 @@ public class ExamController {
 	 */
 	@RequestMapping(path = "/exam/updateExam")
 	public String updateExam(ExamVO examVO, RedirectAttributes redirectAttributes) {
+		int index = 0;
 		
 		try {
 			// 기존 문제의 수정
@@ -202,23 +162,47 @@ public class ExamController {
 			}
 			
 			// 기존문제를 삭제할 경우
-			for(String queId : examVO.getDelqueIdList()) {
-				AnswersheetVO answersheetVO = new AnswersheetVO();
-				answersheetVO.setQueId(queId);
-				
-				QuestionVO questionVO = new QuestionVO();
-				questionVO.setQueId(queId);
-				
-				answersheetService.deleteAnswersheet(answersheetVO);
-				questionService.deleteQuestion(questionVO);
+			if(examVO.getDelqueIdList() != null) {
+				for(String queId : examVO.getDelqueIdList()) {
+					AnswersheetVO answersheetVO = new AnswersheetVO();
+					answersheetVO.setQueId(queId);
+					
+					QuestionVO questionVO = new QuestionVO();
+					questionVO.setQueId(queId);
+					
+					answersheetService.deleteAnswersheet(answersheetVO);
+					questionService.deleteQuestion(questionVO);
+				}
 			}
 			
-		} catch (Exception e) {
+			// 새로운 문제 추가
+			if(examVO.getQueContList() != null) {
+				for (int i=0; i < examVO.getQueContList().size(); i++) {
+					// 시험 아이디를 셋팅
+					QuestionVO questionVO = new QuestionVO();
+					questionVO.setExamId(examVO.getExamId());
+					questionVO.setQueCont(examVO.getQueContList().get(i));
+					questionVO.setQueAnswer(examVO.getQueAnswerList().get(i));
+					questionVO.setQueExplain(examVO.getQueExplainList().get(i));
+					questionVO.setQueScore(examVO.getQueScoreList().get(i));
+					
+					questionService.insertQuestion(questionVO);
+					
+					for (int j = index; j < index + 4; j++) { // 보기가 4개씩 존재
+						// 시험 문제 아이디를 셋팅
+						AnswersheetVO answersheetVO = new AnswersheetVO();
+						answersheetVO.setAnsCont(examVO.getAnsContList().get(j));
+						answersheetVO.setQueId(questionVO.getQueId());
+						
+						answersheetService.insertAnswersheet(answersheetVO);
+					}
+					index =+4;
+				}
+			}
+			
+		}catch (Exception e) {
 			e.printStackTrace();
 		}
-		
-		// 새로운 문제 추가
-		insertExam(examVO);
 		
 		// redirect Attribute
 		redirectAttributes.addFlashAttribute("examVO", examVO);
@@ -277,9 +261,6 @@ public class ExamController {
 		} 
 	}
 
-
-
-
 	/**
 	 * 등록한 시험문제를 삭제하는 메서드
 	 * 
@@ -288,12 +269,15 @@ public class ExamController {
 	 */
 	@RequestMapping(path = "/exam/deleteExam")
 	public String deleteExam(ExamVO examVO, RedirectAttributes redirectAttributes) {
+		AnswersheetVO answersheetVO = new AnswersheetVO();
 		try {
-			for(AnswersheetVO answersheetVO : examVO.getAnswersheetLists()) {
-				answersheetService.deleteAnswersheet(answersheetVO);
-			}
-			for(QuestionVO questionVO : examVO.getQuestionList()) {
-				questionService.deleteQuestion(questionVO);
+			if(examVO.getQuestionList() != null) {
+				for(QuestionVO questionVO : examVO.getQuestionList()) {
+					answersheetVO.setQueId(questionVO.getQueId());
+					answersheetService.deleteAnswersheet(answersheetVO);
+					
+					questionService.deleteQuestion(questionVO);
+				}
 			}
 			examService.deleteExam(examVO);
